refactor(intex): widen HolderBalances to uint32 to match the supply cap
holderBalances cast each balance to uint16 while a holder can accumulate up to the uint32 supply cap across repeated mints and transfers, so a balance above 65535 wrapped around in the view. Widen the struct fields and the casts to uint32. Add a regression test that drives a holder to 80000 and asserts no truncation.

+++ contracts/intex/src/shared/IntexNFT1155.sol
@@ -625,7 +625,7 @@ contract IntexNFT1155 is ERC1155Upgradeable, AccessControlUpgradeable, UUPSUpgra
         uint256 iTok = uint256(seriesId);
         uint256 sTok = _settledTokenId(seriesId);
         return IIntexNFT1155.HolderBalances({
-            issued: uint16(balanceOf(holder, iTok)), settled: uint16(balanceOf(holder, sTok))
+            issued: uint32(balanceOf(holder, iTok)), settled: uint32(balanceOf(holder, sTok))
         });
     }
